+++ create/Template/components/PropUpdateExample/css.js
@@ -1,5 +1,8 @@
 const css = `
-  c-prop-update-example {
+  c-prop-update-example > div {
+    display:flex;
+    flex-direction:column;
+    gap:10px;
   }
 `;