+++ create/Template/components/PropUpdateExample/html.js
@@ -1,8 +1,10 @@
 const html = `
-<h3>Prop Update Example</h3>
-Click the button to set the prop on this element. This will trigger a re-render with the new value.
-<button class="set-data">Set Value on "data-current-time"</button>
-<div class="data-current-time">No Data</div>
+<div>
+    <h3>Prop Update Example</h3>
+    Click the button to set the prop on this element. This will trigger a re-render with the new value.<br/>
+    <button class="set-data">Set Value on "data-current-time"</button><br/>
+    <div class="data-current-time">No Data</div>
+</div>
 `;
 
 export default html;