+++ create/Template/pages/Desktop/css.js
@@ -1,0 +1,6 @@
+const css = `
+  p-desktop {
+  }
+`;
+
+export default css;