+++ create/Template/routes.js
@@ -2,12 +2,16 @@
 import './pages/Error/index.js';
 import './pages/Home/index.js';
 import './pages/About/index.js';
+import './pages/Examples/index.js';
+import './pages/Desktop/index.js';
 
 const routes = {
     '#/error': 'p-error',
     '/': 'p-home',
     '#/home': 'p-home',
-    '#/about': 'p-about'
+    '#/about': 'p-about',
+    '#/examples': 'p-examples',
+    '#/desktop': 'p-desktop'
 };
 
 export { routes };