+++ create/create/api/services.js
@@ -1,7 +1,17 @@
+/**
+ * @module Create-services
+ */
+
 const apps = require('./services/apps');
 const components = require('./services/components');
 const pages = require('./services/pages');
 
+/**
+ * Initializes all services for the specified application.
+ * Calls `start` on apps, components and pages services.
+ *
+ * @param {object} app - express js application instance
+ */
 const start = (app) => {
     apps.start(app);
     components.start(app);